Refactored articles pages
