+++ resources/js/Pages/Articles/Partials/PageHeader.vue
@@ -1,0 +1,7 @@
+<template>
+  <header class="w-full p-4 bg-white">
+    <h1 class="text-xl text-gray-700 font-semibold">
+      {{ $trans('All categories') }}
+    </h1>
+  </header>
+</template>
